requirements.txt
discord.py
requests
random
python.decouple